handling project model updates on homepage
